Creating and Getting TODOs by User

+++ src/app/api/todos/delete-completed/route.ts
@@ -1,11 +1,18 @@
 import { NextResponse } from "next/server";
+import { getUserSession } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
 
 export async function DELETE() {
   try {
+    const user = await getUserSession();
+
+    if (!user)
+      return NextResponse.json({ message: "No Autorizado" }, { status: 401 });
+
     await prisma.todo.deleteMany({
       where: {
         complete: true,
+        userId: user.id,
       },
     });
     return NextResponse.json(
